# openapi.yml
openapi: 3.0.0
info:
  description: API for Stanford Digital Repository indexing service
  version: 1.0.0
  title: Dor Indexing API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://dor-indexing-app-{env}-{node}.stanford.edu'
    description: Production service
    variables:
      env:
        default: prod
      node:
        default: a
        enum: [a b c]
  - url: 'https://dor-indexing-app-{env}-{node}.stanford.edu'
    description: Staging service
    variables:
      env:
        default: stage
      node:
        default: a
        enum: [a b]
  - url: 'https://dor-indexing-app-{env}-{node}.stanford.edu'
    description: Quality Assurance service
    variables:
      env:
        default: qa
      node:
        default: a
        enum: [a b]
tags:
  - name: indexing
    description: Indexing operations
  - name: informational
    description: Informational operations
paths:
  /dor/reindex/{pid}:
    post:
      tags:
        - indexing
      summary: Reindex a repository object, using a network call to retrieve the Cocina model
      description: ''
      operationId: 'dor#reindex'
      parameters:
        - name: pid
          in: path
          description: 'a digital repository identifier'
          required: true
          schema:
            $ref: '#/components/schemas/Druid'
        - name: commitWithin
          in: query
          description: 'time within which to trigger Solr commit'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Object successfully reindexed
          content:
            text/plain: # This is why we can't enable Committee response validation. See TODO in config/application.rb
              schema:
                type: string
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dor/reindex_from_cocina:
    put:
      tags:
        - indexing
      summary: Reindex a repository object using Cocina JSON provided by the caller
      description: ''
      operationId: 'dor#reindex_from_cocina'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required:
                - cocina_object
                - created_at
                - updated_at
              properties:
                cocina_object:
                  description: JSON serialization of the Cocina object.  to be used, created_at and updated_at must also be provided.
                  type: string
                created_at:
                  description: the creation date of the Cocina object
                  type: string
                  format: date-time
                updated_at:
                  description: the most recent modification date of the Cocina object
                  type: string
                  format: date-time
                commitWithin:
                  description: 'time within which to trigger Solr commit'
                  type: integer
      responses:
        '200':
          description: Object successfully reindexed
          content:
            text/plain: # This is why we can't enable Committee response validation. See TODO in config/application.rb
              schema:
                type: string
        '422':
          description: Unprocessable entity (the caller must provide valid Cocina)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dor/delete_from_index/{pid}:
    post:
      tags:
        - indexing
      summary: Remove a repository object's index entry
      description: ''
      operationId: 'dor#delete_from_index'
      parameters:
        - name: pid
          in: path
          description: 'a digital repository identifier'
          required: true
          schema:
            $ref: '#/components/schemas/Druid'
        - name: commitWithin
          in: query
          description: 'time within which to trigger Solr commit'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Object successfully removed from index
          content:
            text/plain: # This is why we can't enable Committee response validation. See TODO in config/application.rb
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dor/queue_size:
    get:
      tags:
        - informational
      summary: Return size of indexing queue
      description: ''
      operationId: 'dor#queue_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    type: integer
components:
  schemas:
    Druid:
      description: Digital Repository Unique Identifier (DRUID) with `druid:` prefix
      type: string
      pattern: '^druid:[b-df-hjkmnp-tv-z]{2}[0-9]{3}[b-df-hjkmnp-tv-z]{2}[0-9]{4}$'
      example: 'druid:bc123df4567'
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    Error:
      type: object
      properties:
        title:
          type: string
          description: 'a short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem.'
          example: Invalid Attribute
        detail:
          type: string
          description: a human-readable explanation specific to this occurrence of the problem.
          example: Title must contain at least three characters.
        source:
          type: object
          properties:
            pointer:
              type: string
              example: /data/attributes/title